+++ extension.js
@@ -95,9 +95,41 @@ function activate(context) {
         }
     });
 
+    let disposableAttributeFormatList = vscode.commands.registerCommand('extension.generateVBAttributesWithFormatList', function () {
+        var editor = vscode.window.activeTextEditor;
+        if (!editor)
+            return; // No open text editor
+
+        var selection = editor.selection;
+        var text = editor.document.getText(selection);
+
+        if (text.length < 1) {
+            vscode.window.showErrorMessage('No selected properties.');
+            return;
+        }
+
+        try {
+            var attrList = createAttributesWithFormatList(text);
+
+            editor.edit(
+                edit => editor.selections.forEach(
+                  selection => 
+                  {
+                    edit.insert(selection.end, attrList);
+                  }
+                )
+              );
+        } 
+        catch (error) {
+            console.log(error);
+            vscode.window.showErrorMessage('Something went wrong! Try that the properties are in this format: "Private p_**** As String"');
+        }
+    });
+
     context.subscriptions.push(disposableGS);
     context.subscriptions.push(disposableConstructors);
     context.subscriptions.push(disposableAttributesList);
+    context.subscriptions.push(createAttributesWithFormatList);
 }
 
 function _transformFirstCharToUpperCase(s) {
@@ -111,19 +143,41 @@ function _extractPropertiesArray(props) {
             type: "",
             attribute: "",
             rawAttribute: "",
+            formatType: "",
+            formatValue: "",
             constStatus: false,
             createStatus: false,
             errorLine: ""
         };
         while (p.startsWith(" ")) p = p.substr(1);
         while (p.startsWith("\t")) p = p.substr(1);
+        while (p.endsWith(" ")) p = p.slice(0, -1); // Remove all extra whitespaces from end of sentence
 
         let words = p.split(" ").map(x => x.replace('\r\n', ''));
         
-        // if words > 7 (i.e ["Private", "Const", "p_*", "As", "Integer", "=", "0"];)
-        if (words.length >= 7) {
+        // if words == 8 (i.e ["Private", "p_*", "As", "Variant", "'", "FORMAT", "NumberFormat", "VALUE", "######.#0"];)
+        if (words.length >= 9) {
+            propObj.type = words[3];
+            propObj.attribute = words[1];
+            
+            propObj.formatType = words[6]
+            propObj.formatValue = words[8]
+
+            propObj.constStatus = false;
+            propObj.createStatus = true;
+        }
+        // if words == 8 (i.e ["p_*", "As", "Variant", "'", "FORMAT", "NumberFormat", "VALUE", "######.#0"];)
+        // if words == 6 (i.e ["Const", "p_*", "As", "Integer", "=", "0"];)
+        else if (words.length == 8 || words.length == 6) {
+            propObj.errorLine = "Declare if the property is Private or Public!";
+        }
+        // if words == 7 (i.e ["Private", "Const", "p_*", "As", "Integer", "=", "0"];)
+        else if (words.length == 7) {
             propObj.type = words[4];
             propObj.attribute = words[2];
+
+            propObj.formatType = ""
+            propObj.formatValue = ""
 
             propObj.constStatus = true;
             propObj.createStatus = true;
@@ -132,6 +186,9 @@ function _extractPropertiesArray(props) {
         else if (words.length == 4) {
             propObj.type = words[3];
             propObj.attribute = words[1];
+
+            propObj.formatType = ""
+            propObj.formatValue = ""
             
             propObj.createStatus = true;            
         }
@@ -139,6 +196,9 @@ function _extractPropertiesArray(props) {
         else if (words.length == 3) {
             propObj.type = words[2];
             propObj.attribute = words[0];
+
+            propObj.formatType = ""
+            propObj.formatValue = ""
             
             propObj.createStatus = true;        
         }
@@ -146,6 +206,9 @@ function _extractPropertiesArray(props) {
         else if (words.length) {
             propObj.type = "Variant";
             propObj.attribute = words[0];
+
+            propObj.formatType = ""
+            propObj.formatValue = ""
 
             propObj.createStatus = true;
         }
@@ -160,7 +223,7 @@ function _extractPropertiesArray(props) {
 
         extProps.push(propObj);
     }
-
+    
     return extProps;
 }
 
@@ -260,6 +323,7 @@ Public Sub Init(`;
                     }
                 }
                 else {
+                    console.error('Error in property:',propList[i])
                     vscode.window.showErrorMessage('Something went wrong! All properties name has to start with a p_! Change ' + propList[i].attribute + ' to p_ATTRIBUTENAME.');
                 }
             }
@@ -308,6 +372,7 @@ Public Sub Init(`;
                     codeBody += codeInnerBody;
                 }
                 else {
+                    console.error('Error in property:',p.attribute)
                     vscode.window.showErrorMessage('Something went wrong! All properties name has to start with a p_! Change ' + p.attribute + ' to p_ATTRIBUTENAME.');
                 }
             }
@@ -367,6 +432,99 @@ End Sub`;
             }
             else {
                 vscode.window.showErrorMessage('Something went wrong! All properties name has to start with a p_! Change ' + propList[i].attribute + ' to p_ATTRIBUTENAME.');
+            }
+        }
+    }
+
+    return generatedCode;
+}
+
+function createAttributesWithFormatList(propsText) {
+    var properties = propsText.split('\r\n').filter(x => x.length > 2);
+    var propList = _extractPropertiesArray(properties);
+
+    var generatedCode = 
+`
+`;
+
+    var errorMessage = `Something went wrong (fix them before proceeding) in the following lines: `;
+    var errorList = propList.filter(p => !p.createStatus);
+
+    if (errorList.length) {
+        for(let i=0; i < errorList.length; i++) {
+            errorMessage += `\n${errorList[i].errorLine}`;
+        }
+        vscode.window.showErrorMessage(errorMessage);
+    }
+    else {
+        generatedCode += 
+`Private p_attributesList() As Variant
+Private p_attributesFormatTypesList() As Variant
+Private p_attributesFormatValuesList() As Variant
+
+Private Sub class_initialize()
+\tp_attributesList = Array(`;
+
+        for (let i = 0, j = 0; i < propList.length; i++, j++) {
+            let rawAttribute = propList[i].attribute.split('_')[1]
+            if (rawAttribute) {
+                generatedCode += `"${rawAttribute}"`;
+                if (i !== propList.length - 1) {
+                    generatedCode += ', ';
+                    // Breaking lines (VB editor in excel has character limits)
+                    if (j == 5) {
+                        j = 0;
+                        generatedCode +=
+` _ 
+`;
+                    }
+                }
+                else {
+                    generatedCode += `)`;
+                }
+            }
+        }
+
+        generatedCode += 
+`
+\tp_attributesFormatTypesList = Array(`;
+
+        for (let i = 0, j = 0; i < propList.length; i++, j++) {
+            generatedCode += `"${propList[i].formatType}"`;
+            if (i !== propList.length - 1) {
+                generatedCode += ', ';
+                // Breaking lines (VB editor in excel has character limits)
+                if (j == 5) {
+                    j = 0;
+                    generatedCode +=
+` _ 
+`;
+                }
+            }
+            else {
+                generatedCode += `)`;
+            }
+        }
+
+        generatedCode += 
+`
+\tp_attributesFormatValuesList = Array(`;
+
+        for (let i = 0, j = 0; i < propList.length; i++, j++) {
+            generatedCode += `"${propList[i].formatValue}"`;
+            if (i !== propList.length - 1) {
+                generatedCode += ', ';
+                // Breaking lines (VB editor in excel has character limits)
+                if (j == 5) {
+                    j = 0;
+                    generatedCode +=
+` _ 
+`;
+                }
+            }
+            else {
+                generatedCode += `)
+End Sub`;
             }
         }
     }